Move table.Row class into its own submodule table/row.py
